fix: modified entities and usecases, added UserRepositoryService

+++ core/src/main/java/ru/amorozov/domain/entities/TaskStatus.java
@@ -1,13 +1,12 @@
 package ru.amorozov.domain.entities;
 
-
 import ru.amorozov.domain.entities.enums.Status;
 
 /**
  * @author Aleksey Morozov
  * @since 25.08.2022
  */
-public record TaskStatus(int id,
+public record TaskStatus(long id,
                          Status status) implements BasicEntity {
 
 }
