+++ core/src/main/java/ru/amorozov/domain/entities/UserRole.java
@@ -7,5 +7,6 @@ import ru.amorozov.domain.entities.enums.Role;
  * @author Aleksey Morozov
  * @since 26.08.2022
  */
-public record UserRole(int id, @NonNull Role role) implements BasicEntity {
+public record UserRole(long id,
+                       @NonNull Role role) implements BasicEntity {
 }
